Refactor evaluation template package resolution
